requirements.txt
python==3.13.7

google-genai==1.59.0
streamlit==1.49.1
